+++ tests/unit/test_plugin.py
@@ -367,7 +367,9 @@ def test_pytest_addoption_adds_correct_ini_file_arguments():
         'rp_api_retries',
         'rp_skip_connection_test',
         'rp_launch_timeout',
-        'rp_client_type'
+        'rp_client_type',
+        'rp_connect_timeout',
+        'rp_read_timeout'
     )
     mock_parser = mock.MagicMock(spec=Parser)